+++ src/server/ws-routes/initializeWebSocketRouting.js
@@ -7,6 +7,6 @@ module.exports = initializeWebSocketRouting
 
 function initializeWebSocketRouting(io, protocol = require('../protocol/protocol.json')) {
 
-	Utilities.createWebsocketRoute(io, protocol.room, require('./services/Room'), new ConnectionsRepository())
+	Utilities.createWebsocketRoute(io, protocol.room, require('./controllers/Room'), new ConnectionsRepository())
 
 }
